+++ src/psqt.cpp
@@ -1,6 +1,4 @@
 #include <map>
-#include <cstdint>
-
 #include "psqt.h"
 
 int piece_values[2][6] = { { 126, 781, 825, 1276, 2538, 0}, { 208, 854, 915, 1380,  2682, 0} };
@@ -28,8 +26,8 @@ int w_rook_eg[64]   = { -9, -13, -10, -9, -9, -10, -13, -9, -12, -9, -1, -2, -2,
 int b_rook_eg[64]   = { 18, 0, 19, 13, 13, 19, 0, 18, 4, 5, 20, -5, -5, 20, 5, 4, 6, 1, -7, 10, 10, -7, 1, 6, -5, 8, 7, -6, -6, 7, 8, -5, -6, 1, -9, 7, 7, -9, 1, -6, 6, -8, -2, -6, -6, -2, -8, 6, -12, -9, -1, -2, -2, -1, -9, -12, -9, -13, -10, -9, -9, -10, -13, -9 };
 int w_queen_eg[64]  = { -69, -57, -47, -26, -26, -47, -57, -69, -54, -31, -22, -4, -4, -22, -31, -54, -39, -18, -9, 3, 3, -9, -18, -39, -23, -3, 13, 24, 24, 13, -3, -23, -29, -6, 9, 21, 21, 9, -6, -29, -38, -18, -11, 1, 1, -11, -18, -38, -50, -27, -24, -8, -8, -24, -27, -50, -74, -52, -43, -34, -34, -43, -52, -74 };
 int b_queen_eg[64]  = { -74, -52, -43, -34, -34, -43, -52, -74, -50, -27, -24, -8, -8, -24, -27, -50, -38, -18, -11, 1, 1, -11, -18, -38, -29, -6, 9, 21, 21, 9, -6, -29, -23, -3, 13, 24, 24, 13, -3, -23, -39, -18, -9, 3, 3, -9, -18, -39, -54, -31, -22, -4, -4, -22, -31, -54, -69, -57, -47, -26, -26, -47, -57, -69 };
-int w_king_eg[64]   = { -65, 23, 16, -15, -56, -34, 2, 13, 29, -1, -20, -7, -8, -4, -38, -29, -9, 24, 2, -16, -20, 6, 22, -22, -17, -20, -12, -27, -30, -25, -14, -36, -49, -1, -27, -39, -46, -44, -33, -51, -14, -14, -22, -46, -44, -30, -15, -27, 1, 7, -8, -64, -43, -16, 9, 8, -15, 36, 12, -54, 8, -28, 24, 14 };
-int b_king_eg[64]   = { -15, 36, 12, -54, 8, -28, 24, 14, 1, 7, -8, -64, -43, -16, 9, 8, -14, -14, -22, -46, -44, -30, -15, -27, -49, -1, -27, -39, -46, -44, -33, -51, -17, -20, -12, -27, -30, -25, -14, -36, -9, 24, 2, -16, -20, 6, 22, -22, 29, -1, -20, -7, -8, -4, -38, -29, -65, 23, 16, -15, -56, -34, 2, 13};
+int w_king_eg[64]   = { 1, 45, 85, 76, 76, 85, 45, 1, 53, 100, 133, 135, 135, 133, 100, 53, 88, 130, 169, 175, 175, 169, 130, 88, 103, 156, 172, 172, 172, 172, 156, 103, 96, 166, 199, 199, 199, 199, 166, 96, 92, 172, 184, 191, 191, 184, 172, 92, 47, 121, 116, 131, 131, 116, 121, 47, 11, 59, 73, 78, 78, 73, 59, 11 };
+int b_king_eg[64]   = { 11, 59, 73, 78, 78, 73, 59, 11, 47, 121, 116, 131, 131, 116, 121, 47, 92, 172, 184, 191, 191, 184, 172, 92, 96, 166, 199, 199, 199, 199, 166, 96, 103, 156, 172, 172, 172, 172, 156, 103, 88, 130, 169, 175, 175, 169, 130, 88, 53, 100, 133, 135, 135, 133, 100, 53, 1, 45, 85, 76, 76, 85, 45, 1 };
 
 std::map<int, int*> piece_to_mg =
 {
@@ -61,15 +59,4 @@ std::map<int, int*> piece_to_eg =
 { 9,   b_rook_eg },
 { 10,  b_queen_eg },
 { 11,  b_king_eg },
-};
-
-uint8_t arrCenterManhattanDistance[64] = { // char is sufficient as well, also unsigned
-  6, 5, 4, 3, 3, 4, 5, 6,
-  5, 4, 3, 2, 2, 3, 4, 5,
-  4, 3, 2, 1, 1, 2, 3, 4,
-  3, 2, 1, 0, 0, 1, 2, 3,
-  3, 2, 1, 0, 0, 1, 2, 3,
-  4, 3, 2, 1, 1, 2, 3, 4,
-  5, 4, 3, 2, 2, 3, 4, 5,
-  6, 5, 4, 3, 3, 4, 5, 6
 };
