+++ src/psqt.h
@@ -32,5 +32,3 @@ extern int b_king_eg[64];
 extern std::map<int, int*> piece_to_mg;
 
 extern std::map<int, int*> piece_to_eg;
-
-extern uint8_t arrCenterManhattanDistance[64];
